2375 update rules round2 (#704)
* fixed work location rule

* added more lineage breaks

* regex pattern updates

* updated doc strings and rules

* regex pattern changes

* a dummy test change

* regex updates

* fixed misspecified bool condition

* updated tests

+++ tests/pipeline/test_reclassify_work_variables.py
@@ -36,7 +36,7 @@ def test_reclassify_work_variables(spark_session, load_test_cases):
 
     input_schema = t.StructType(
         [
-            t.StructField("row_id", t.IntegerType()),
+            t.StructField("row_id", t.FloatType()),
             t.StructField("rule", t.StringType()),
             t.StructField("work_main_job_title", t.StringType()),
             t.StructField("work_main_job_role", t.StringType()),
